# Fix flicker