--- a/scenario_gym/__init__.py
+++ b/scenario_gym/__init__.py
@@ -50,4 +50,4 @@ __all__ = [
     "xosc_interface",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"