--- a/setup.py
+++ b/setup.py
@@ -54,5 +54,5 @@ setup(
         where=".",
         include=["scenario_gym", "scenario_gym.*"],
     ),
-    version="0.1.0",
+    version="0.2.0",
 )