--- a/tests/test_scenario.py
+++ b/tests/test_scenario.py
@@ -44,7 +44,7 @@ def test_import_scenario(scenario_path):
     assert all((e.catalog_entry.catalog_category is not None for e in s.entities))
     assert s.road_network is not None
 
-    s = import_scenario(scenario_path, relabel=True)
+    s = import_scenario(scenario_path)
     e1, e2 = s.entities[1], s.entities[2]
     assert e1.ref[:-2] == e1.catalog_entry.catalog_category.lower()
     assert e2.ref[:-2] == e2.catalog_entry.catalog_category.lower()