This PR fixes the flickering effect when rendering driveable surface boundaries. This was caused by caching of geometries coordinates. Geometries were cached using `id(geom)` which works fine for geometries fixed in memory e.g. the boundary and centre lines. For geometries accessed as properties of other shapely geometries such as interiors of polygons the id changes each time the geometry is queried meaning the cache failed and would often be replaced by another object.

To fix this I have just removed the caching where it uses a non-fixed geometry. This comes with a considerable performance loss:

On this branch:
```
$ pytest tests/test_speeds.py -s -k test_render_speed --speed_tests
tests/test_speeds.py Completed in 20.88s per scenario, 3.185e+04μs per step.
```

On main:
```
$ pytest tests/test_speeds.py -s -k test_render_speed --speed_tests
tests/test_speeds.py Completed in 4.248s per scenario, 6.479e+03μs per step.
```
This might be fixed by Shapely 2.0 as then geometries will be hashable however this is only in alpha release. I will investigate a different solution in the meantime.